// DXC_Frontend/src/api/models/hotlineDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface HotlineDto {
  publicId?: string
  categoryPublicId?: string
  /** @nullable */
  categoryName: string | null
  /** @nullable */
  phoneNumber: string | null
  /** @nullable */
  contactName: string | null
  /** @nullable */
  description?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}